Significant content update.
Other than the graphics (which are obviously largely different) the
biggest changes here come from the page background being set to the
game background, the phaser frame is now transparent, and the layouts
becoming responsive.  I also refactored the code to exist within app.js
to make the page scripts a little more… gooder? The main menu is done,
the game screen is in progress (but I’m about to leave for the day
hence the push) but the game over screen has not been touched.
Learning a lot of cool tricks about responsive phaser games from
netexl.com